# Documents/SUNNU_MBAYE_MI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/ajoutProduit": {
            "post": {
                "summary": "Ajouter un produit",
                "operationId": "f19499da27676744a66014051960bc8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom_produit": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "quantite": {
                                        "type": "integer"
                                    },
                                    "prix": {
                                        "type": "integer"
                                    },
                                    "categorie_id": {
                                        "type": "integer"
                                    },
                                    "images": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "produit": {
                                            "$ref": "#/components/schemas/Produit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton invalide"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/updateproduit/{id}": {
            "put": {
                "summary": "Mettre à jour un produit",
                "operationId": "558bbbb3524de96b7c4de056a96feb48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "nom_produit": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "quantite": {
                                        "type": "integer"
                                    },
                                    "prix": {
                                        "type": "integer"
                                    },
                                    "categorie_id": {
                                        "type": "integer"
                                    },
                                    "images": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "produit": {
                                            "$ref": "#/components/schemas/Produit"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé - Jeton invalide"
                    },
                    "403": {
                        "description": "Accès refusé - Utilisateur non autorisé à modifier cette annonce"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/supProduit/{id}": {
            "delete": {
                "summary": "Supprime un produit",
                "operationId": "ed54e9c91d63de3bfbdfc1cac02787f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du produit à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit supprimé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Produit supprimé avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Interdit d'accès"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/produitrecherche": {
            "post": {
                "summary": "Recherche de produits",
                "operationId": "d70a7b3336cf183d279259a55e345a70",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produits trouvés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "type": "array"
                                        },
                                        "item": {
                                            "type": "string"
                                        },
                                        "categories": {
                                            "type": "array"
                                        },
                                        "newProduct": {
                                            "type": "array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/rechercheproduit": {
            "post": {
                "summary": "Recherche de produits",
                "operationId": "e9be2194dfbdda4935cfae7c8a690182",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produits trouvés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "products": {
                                            "type": "array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/api/filtrer-produits-par-categorie": {
            "post": {
                "summary": "Filtrer les produits par catégorie",
                "description": "Filtrer les produits par catégorie.",
                "operationId": "2b802c098d1020bf93d3f02a43ce6c66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "categorie_id"
                                ],
                                "properties": {
                                    "categorie_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Renvoie les produits filtrés par catégorie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "produits": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Produit"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non autorisé"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}